feat(core): Phase 5.11 - slim __init__.py and extract canonical _profile_print module
